Create calculator for pre-departure factor and create generic top delay calculator

+++ src/main/java/com/flightDelay/flightdelayapi/statisticsFactors/calculator/AverageFactorCalculator.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface AverageFactorCalculator {
 
-    double calculateAverage(List<Integer> numerator, List<Integer> denominator);
+    double calculateAverage(List<Double> numerator, List<Double> denominator);
 
     double calculateAverage(double numerator, double denominator);
 }
